+++ client/projects/client/src/app/modules/main/main-routing.module.ts
@@ -2,6 +2,7 @@ import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { MainComponent } from './main.component';
 import { SettingsComponent } from './components/settings/settings.component';
+import { RoomComponent } from './components/room/room.component';
 
 const routes: Routes = [
     {
@@ -11,6 +12,10 @@ const routes: Routes = [
             {
                 path: "settings",
                 component: SettingsComponent
+            },
+            {
+                path: "room/:jid",
+                component: RoomComponent
             },
         ]
     }