+++ client/projects/client/src/app/store/chat/reducer.ts
@@ -1,0 +1,10 @@
+import { combineReducers } from '@ngrx/store';
+import { roomsReducer, RoomsState } from './reducers/rooms';
+
+export interface XmppState {
+    rooms: RoomsState,
+}
+
+export const xmppReducer = combineReducers({
+    rooms: roomsReducer
+});