+++ client/projects/client/src/app/store/chat/reducers/def.ts
@@ -1,0 +1,6 @@
+import { DiscoInfo } from "stanza/protocol"
+
+export interface Room {
+    id: string
+    info: DiscoInfo
+}